refactor(#8): improve instruction handling and disassembly methods

+++ src/core/mod.rs
@@ -39,7 +39,7 @@ impl Dmg {
             match self.cpu.exec(opcode) {
                 Ok(effect) => {
                     self.cpu.cycles += effect.cycles as usize;
-                    self.cpu.pc += effect.len;
+                    self.cpu.pc += effect.len as u16;
                     match effect.flags {
                         Some(f) => self.cpu.f |= f,
                         None => {}
